requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4
flask_cors
python-dotenv
psycopg2-binary==2.9.10
bcrypt
cloudinary